style: eliminar comentarios adicionales y organizar el codigo evitando redundancias
