+++ frontend/src/app/(auth)/register/page.tsx
@@ -1,19 +1,12 @@
 'use client';
 
 import Logo from '@/ui/logo';
-import { Input, Button } from '@nextui-org/react';
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { Icon } from '@iconify-icon/react';
 import { useRouter } from 'next/navigation';
+import { PasswordInputProps } from '@/types/auth';
 
-type PasswordInputProps = {
-  label: string;
-  placeholder?: string;
-  isRequired?: boolean;
-  name: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  error?: string;
-};
+import { Icon } from '@iconify-icon/react';
+import { Input, Button } from '@nextui-org/react';
 
 function PasswordInput({
   label,
