+++ frontend/src/types/layout.ts
@@ -1,0 +1,8 @@
+export interface NavbarProps {
+  title: string;
+}
+
+export interface AvatarImgProps  {
+  src: string;
+  isLoading?: boolean;
+};
