+++ frontend/src/ui/navbar.tsx
@@ -2,6 +2,7 @@
 
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
+
 import { Icon } from '@iconify-icon/react';
 
 export default function Navbar() {
