+++ frontend/src/ui/postCard.tsx
@@ -1,13 +1,14 @@
 'use client';
 
+import Link from 'next/link';
 import { Icon } from '@iconify-icon/react';
 import { Avatar, Divider } from '@nextui-org/react';
+import { PostCardProps } from '@/types/posts';
+
 import PostActions from '@/ui/postActions';
-import Link from 'next/link';
-import { PostCardProps } from '@/types';
 
 export default function PostCard({ frase, onLike }: PostCardProps) {
-  const usuario = {
+  const baseUser = {
     name: frase.usuarioId?.name || 'Bitacora UP',
     username: frase.usuarioId?.username || 'bitacoraup',
     avatar: frase.usuarioId?.avatar || 'https://i.ibb.co/ZNyjQ2g/favicon.jpg',
@@ -16,25 +17,23 @@ export default function PostCard({ frase, onLike }: PostCardProps) {
   return (
     <>
       <div className="flex items-start space-x-4 p-4">
-        {/* Avatar del usuario */}
         <Avatar
           size="md"
           className="shadow-lg"
           name={frase.usuarioId?.avatar}
-          src={usuario.avatar}
+          src={baseUser.avatar}
         />
 
-        {/* Contenido del post */}
         <div className="flex-1">
           <div className="flex items-center">
-            <p className="font-medium ">{usuario.name}</p>
+            <p className="font-medium ">{baseUser.name}</p>
             <Icon
               icon="bitcoin-icons:verify-filled"
               width="16"
               className="text-blue-400 align-middle"
             />
           </div>
-          <p className="text-sm text-gray-500">@{usuario.username}</p>
+          <p className="text-sm text-gray-500">@{baseUser.username}</p>
           <div className="p-4 pt-2 mt-2 rounded-md border border-gray-300/70">
             <Link href={`/post/${frase._id}`}>
               <p>{frase.frase}</p>
@@ -48,7 +47,7 @@ export default function PostCard({ frase, onLike }: PostCardProps) {
               <p className="mt-4 text-sm text-gray-500">- {frase.autor}</p>
             </Link>
           </div>
-          {/* Botones de interacción */}
+
           <PostActions
             likes={frase.likes}
             comments={frase.comentarios?.length || 0}
